requirements.txt
playwright==1.41.1
pytest==8.4.1
pytest-playwright==0.4.3
locust==2.38.1
python-dotenv==1.0.0
google-genai==1.29.0
